fixed linux daemons
